+++ base/src/main/java/eu/fraho/spring/securityJwt/service/JwtTokenService.java
@@ -12,7 +12,6 @@ import eu.fraho.spring.securityJwt.dto.AccessToken;
 import eu.fraho.spring.securityJwt.dto.JwtUser;
 import eu.fraho.spring.securityJwt.dto.RefreshToken;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
@@ -25,6 +24,7 @@ public interface JwtTokenService {
      * See {@link JwtUser#applyClaims(JWTClaimsSet)} for a list of copied / supported attributes.
      *
      * @param token The JWT to parse
+     * @param <T>   Any configured subtype of JwtUser
      * @return an user object
      */
     <T extends JwtUser> Optional<T> parseUser(@NotNull String token);
@@ -34,6 +34,7 @@ public interface JwtTokenService {
      * See {@link JwtUser#toClaims()} for a list of copied / supported attributes.
      *
      * @param user The user to create the token for.
+     * @param <T>  Any configured subtype of JwtUser
      * @return The token (Base64 encoded)
      * @throws JOSEException            When the token could not be signed
      * @throws IllegalArgumentException If no private key is specified (service cannot sign tokens)
@@ -65,13 +66,6 @@ public interface JwtTokenService {
     Optional<String> getToken(@NotNull HttpServletRequest request);
 
     /**
-     * @return The validity of newly created access tokens in seconds
-     */
-    @NotNull
-    @Deprecated
-    Integer getExpiration();
-
-    /**
      * Gracefully ask this service if refresh token support is enabled
      * by a third-party addon.
      *
@@ -88,20 +82,7 @@ public interface JwtTokenService {
      * @return The refresh token (Base64 encoded) or {@code null} if property {@code fraho.jwt.refresh.enabled} is set to {@code false}.
      */
     @NotNull
-    RefreshToken generateRefreshToken(@NotNull String user);
-
-    /**
-     * Generate a simple refresh token. This token is not a "normal" JWT.
-     * It's just a base64-encoded string of random data.
-     * Returns {@code null} if the property {@code fraho.jwt.refresh.enabled} is set to {@code false}.
-     *
-     * @param user     The refresh token will be issued to this user
-     * @param deviceId device id for this refresh token to support multiple tokens / devices.
-     *                 This id will be truncated to the configured length.
-     * @return The refresh token (Base64 encoded) or {@code null} if property {@code fraho.jwt.refresh.enabled} is set to {@code false}.
-     */
-    @NotNull
-    RefreshToken generateRefreshToken(@NotNull String user, @Nullable String deviceId);
+    RefreshToken generateRefreshToken(JwtUser user);
 
     /**
      * Use and invalidate a refresh token for the given user using the default device id.
@@ -109,11 +90,11 @@ public interface JwtTokenService {
      * Returns always {@code false} if property {@code fraho.jwt.refresh.enabled} is set to {@code false}
      * or no private key for signing is configured.
      *
-     * @param username     the user requesting a new access token
      * @param refreshToken the refresh token to use (base64 encoded)
+     * @param <T>          Any configured subtype of JwtUser
      * @return {@code true} only if the refresh token is valid and not been used before
      */
-    boolean useRefreshToken(@NotNull String username, @NotNull String refreshToken);
+    <T extends JwtUser> Optional<T> useRefreshToken(@NotNull String refreshToken);
 
     /**
      * Use and invalidate a refresh token for the given user.
@@ -121,25 +102,11 @@ public interface JwtTokenService {
      * Returns always {@code false} if property {@code fraho.jwt.refresh.enabled} is set to {@code false}
      * or no private key for signing is configured.
      *
-     * @param username     the user requesting a new access token
-     * @param deviceId     Optional device id for this refresh token to support multiple tokens / devices
-     *                     This id will be truncated to the configured length.
-     * @param refreshToken the refresh token to use (base64 encoded)
+     * @param token the refresh token to use
+     * @param <T>   Any configured subtype of JwtUser
      * @return {@code true} only if the refresh token is valid and not been used before
      */
-    boolean useRefreshToken(@NotNull String username, @Nullable String deviceId, @NotNull String refreshToken);
-
-    /**
-     * Use and invalidate a refresh token for the given user.
-     * Each refresh token may only be used once.
-     * Returns always {@code false} if property {@code fraho.jwt.refresh.enabled} is set to {@code false}
-     * or no private key for signing is configured.
-     *
-     * @param username the user requesting a new access token
-     * @param token    the refresh token to use
-     * @return {@code true} only if the refresh token is valid and not been used before
-     */
-    boolean useRefreshToken(@NotNull String username, @NotNull RefreshToken token);
+    <T extends JwtUser> Optional<T> useRefreshToken(@NotNull RefreshToken token);
 
     /**
      * List all currently active refresh tokens from the backend cache.
@@ -148,42 +115,40 @@ public interface JwtTokenService {
      * The {@code key} is the username, the {@code value} is the list of tokens for that user
      */
     @NotNull
-    Map<String, List<RefreshToken>> listRefreshTokens();
+    Map<Long, List<RefreshToken>> listRefreshTokens();
 
     /**
      * List all currently active refresh tokens for the given username.
      *
-     * @param username the username to filter the refresh tokens
+     * @param user the user to filter the refresh tokens
      * @return list of refresh tokens
      */
     @NotNull
-    List<RefreshToken> listRefreshTokens(@NotNull String username);
+    List<RefreshToken> listRefreshTokens(@NotNull JwtUser user);
 
     /**
      * Revoke the given token from the stored refresh token list
      *
-     * @param username username
-     * @param token    the token to remove
+     * @param token the token to remove
      * @return {@code true} if the token was revoked
      */
-    boolean revokeRefreshToken(@NotNull String username, @NotNull RefreshToken token);
+    boolean revokeRefreshToken(@NotNull RefreshToken token);
 
     /**
      * Revoke the given token from the stored refresh token list
      *
-     * @param username username
-     * @param deviceId the deviceId to remove
+     * @param token The token to revoke
      * @return {@code true} if the token was revoked
      */
-    boolean revokeRefreshToken(@NotNull String username, @Nullable String deviceId);
+    boolean revokeRefreshToken(@NotNull String token);
 
     /**
      * Revokes all  refresh tokens from the stored refresh token list for the specified user
      *
-     * @param username username
+     * @param user The user which tokens should be revoked
      * @return count of revoked tokens
      */
-    int revokeRefreshTokens(@NotNull String username);
+    int revokeRefreshTokens(@NotNull JwtUser user);
 
     /**
      * Revoke all tokens from the stored refresh token list