+++ base/src/test/java/eu/fraho/spring/securityJwt/it/spring/UserDetailsServiceTestImpl.java
@@ -7,6 +7,8 @@
 package eu.fraho.spring.securityJwt.it.spring;
 
 import eu.fraho.spring.securityJwt.dto.JwtUser;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.ObjectFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -19,15 +21,16 @@ import org.springframework.stereotype.Service;
 import java.util.Collections;
 
 @Service
+@RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class UserDetailsServiceTestImpl implements UserDetailsService {
     public static final String BASE32_TOTP = "MZXW6YTBOI======";
 
     private int noRefreshAccessCount = 0;
 
-    @Autowired
+    @NonNull
     private PasswordEncoder passwordEncoder;
 
-    @Autowired
+    @NonNull
     private ObjectFactory<JwtUser> jwtUser;
 
     @Override