+++ hibernate-spring-boot-starter/src/main/java/eu/fraho/spring/securityJwt/starter/SecurityJwtHibernateAutoConfiguration.java
@@ -16,6 +16,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.core.userdetails.UserDetailsService;
 
 @Configuration
 @AutoConfigureAfter(SecurityJwtBaseAutoConfiguration.class)
@@ -25,8 +26,9 @@ import org.springframework.context.annotation.Configuration;
 public class SecurityJwtHibernateAutoConfiguration {
     @Bean
     @ConditionalOnMissingBean
-    public RefreshTokenStore refreshTokenStore(final JwtRefreshConfiguration jwtRefreshConfiguration) {
+    public RefreshTokenStore refreshTokenStore(final JwtRefreshConfiguration jwtRefreshConfiguration,
+                                               final UserDetailsService userDetailsService) {
         log.debug("Register HibernateTokenStore");
-        return new HibernateTokenStore(jwtRefreshConfiguration);
+        return new HibernateTokenStore(jwtRefreshConfiguration, userDetailsService);
     }
 }